[COR-502] break out chunker options and add tests for validation
